requirements.txt
openmeteo_requests
requests-cache
pandas
retry_requests
pillow
numpy
bleak
playwright